More refactoring and introduced a heartbeat into the worker.
It is a little odd using CherryPy and Flask, but here is the reasoning.

  1. CherryPy is a stable, configurable and reliable server.
  2. It has hooks to start background threads to perform tasks as
     necessary in process. (ie the heartbeat)
  3. It has good process support for signals

The expectation is that we would never have 1000s of workers, which
would make the heartbeats a potential bottleneck if they are all
constantly pinging the master. The heartbeat can also be configurable to
be slower or turned off completely. It will still register on start up
and the master will delete it if it can't reach the host.

There is still a potential for a race condition where the master finds a
host, and the host shuts down prior to actually starting the process,
but I think we can handle that gracefully and just realize the POST
request failed and find a new host.

+++ dad/master/utils.py
@@ -28,9 +28,12 @@ def find_host():
         if num_procs >= procs:
             current_host = host
 
+    if not current_host:
+        raise Exception('No workers available!')
+
     # See if the host is up and running. If not let's delete it.
     try:
-        requests.get('http://%s/' % str(host))
+        requests.get('http://%s/' % str(current_host))
         return current_host
     except requests.exceptions.ConnectionError:
         db.session.delete(host)
